Refactor how files are handled
