+++ app/api/octopub/datasets/files/update.rb
@@ -1,0 +1,51 @@
+module Octopub
+  module Datasets
+    module Files
+      class Update < Grape::API
+
+        before do
+          authenticate!
+          find_dataset
+        end
+
+        namespace :datasets do
+
+          namespace ':id' do
+            namespace :files do
+              desc 'Updates a file or files in an existing dataset'
+              params do
+                requires :id, type: Integer, desc: 'The ID of the dataset'
+                requires :file_id, type: Integer, desc: 'The ID of the file'
+                requires :file, type: Hash do
+                  optional :description, type: String, desc: 'A short description of the file'
+                  optional :file, type: File, desc: 'The actual file'
+                end
+              end
+              put ':file_id' do
+                params.files = [
+                  {
+                    id: params.file_id,
+                    description: params.file.description,
+                    file: params.file.file
+                  }
+                ]
+
+                process_files(params.files)
+
+                job = UpdateDataset.perform_async(@dataset.id, current_user.id, {}, params.files)
+
+                status 202
+                {
+                  job_url: api_jobs_path(id: job)
+                }
+              end
+
+            end
+          end
+
+        end
+
+      end
+    end
+  end
+end
