+++ app/api/octopub/datasets/update.rb
@@ -28,58 +28,6 @@ module Octopub
           }
         end
 
-        namespace ':id' do
-          desc 'Adds a file or files to an existing dataset'
-          params do
-            requires :id, type: Integer, desc: 'The ID of the dataset'
-            requires :files, type: Array do
-              requires :title, type: String, desc: 'The name of the file'
-              requires :description, type: String, desc: 'A short description of the file'
-              requires :file, type: File, desc: 'The actual file'
-            end
-          end
-          post :files do
-            process_files(params.files) if params.files
-
-            job = UpdateDataset.perform_async(@dataset.id, current_user.id, {}, params["files"])
-
-            status 202
-            {
-              job_url: api_jobs_path(id: job)
-            }
-          end
-
-          namespace :files do
-            desc 'Updates a file or files in an existing dataset'
-            params do
-              requires :id, type: Integer, desc: 'The ID of the dataset'
-              requires :file_id, type: Integer, desc: 'The ID of the file'
-              requires :file, type: Hash do
-                optional :description, type: String, desc: 'A short description of the file'
-                optional :file, type: File, desc: 'The actual file'
-              end
-            end
-            put ':file_id' do
-              params.files = [
-                {
-                  id: params.file_id,
-                  description: params.file.description,
-                  file: params.file.file
-                }
-              ]
-
-              process_files(params.files)
-
-              job = UpdateDataset.perform_async(@dataset.id, current_user.id, {}, params.files)
-
-              status 202
-              {
-                job_url: api_jobs_path(id: job)
-              }
-            end
-
-          end
-        end
       end
 
     end
